requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
aiomysql>=0.2.0
asyncio
